Upgrade vue and nuxt
